Move recipe stuff up in package hierarchy
